Backtesting with marginal requirement
So far, the code considersmarginal requirements. However, we have a problem with infesibility. Some assets become illiquid during the investment period. If we heavily long those assets, we would face a problem to sell them for cash. Next we would try allowing an investor to put extra cash in the portfolio. However, there will be a penalty to the amount of extra cash used.